requirements.txt
PyYAML>=6.0
pytest>=8.0
groq>=0.21.0
GitPython>=3.1.43
networkx>=3.4
pydantic>=2.8
openapi-schema-pydantic>=1.2.4
tree-sitter>=0.23.2
